refactor: audio manager
