+++ Clairfy/Features/HomeScreen/VoiceRecordingVC.swift
@@ -2,18 +2,6 @@ import UIKit
 import AVFoundation
 
 class VoiceRecordingViewController: UIViewController, AVAudioRecorderDelegate {
-    // MARK: Properties
-    private var timer: Timer?
-    private var elapsedTime: TimeInterval = 0
-    private var recordingState: RecordingState = .stopped
-    var recordingAnimationTimer: Timer?
-    let recordingImages = [UIImage(named: "rec1"), UIImage(named: "rec3")]
-    var currentImageIndex = 0
-    private var audioRecorder: AVAudioRecorder?
-    private var audioURL: URL?
-    private var recordings: [URL] = []
-    private var audioID: UUID?
-
     // MARK: Subviews
     lazy var recordingImage: UIImageView = {
         var imageView = UIImageView()
@@ -75,18 +63,10 @@ class VoiceRecordingViewController: UIViewController, AVAudioRecorderDelegate {
             let playImage = UIImage(systemName: "trash.fill", withConfiguration: symbolConfig)
             button.setImage(playImage, for: .normal)
         }
-        
-        /// opcional: padding
-        //button.contentEdgeInsets = UIEdgeInsets(top: 0, left: 5, bottom: 0, right: 0)
-        
-        /// constraints
-        NSLayoutConstraint.activate([
-            button.widthAnchor.constraint(equalToConstant: 56),
-            button.heightAnchor.constraint(equalToConstant: 56)
-        ])
-        
+
         return button
     }()
+    
     lazy var finishedAudioButton: UIButton = {
         var button = UIButton()
         button.translatesAutoresizingMaskIntoConstraints = false
@@ -102,18 +82,10 @@ class VoiceRecordingViewController: UIViewController, AVAudioRecorderDelegate {
             let playImage = UIImage(systemName: "checkmark", withConfiguration: symbolConfig)
             button.setImage(playImage, for: .normal)
         }
-        
-        /// opcional: padding
-        //button.contentEdgeInsets = UIEdgeInsets(top: 0, left: 5, bottom: 0, right: 0)
-        
-        /// constraints
-        NSLayoutConstraint.activate([
-            button.widthAnchor.constraint(equalToConstant: 56),
-            button.heightAnchor.constraint(equalToConstant: 56)
-        ])
-        
+
         return button
     }()
+    
     lazy var buttonsStackView: UIStackView = {
         let stackView = UIStackView(arrangedSubviews: [deleteAudioButton, recordButton, finishedAudioButton])
         stackView.axis = .horizontal
@@ -122,6 +94,11 @@ class VoiceRecordingViewController: UIViewController, AVAudioRecorderDelegate {
         stackView.spacing = 40
         stackView.translatesAutoresizingMaskIntoConstraints = false
         return stackView
+    }()
+    
+    // MARK: Proprieties
+    private lazy var recorder: AudioRecordManager = {
+        return AudioRecordManager(voiceRecordVC: self)
     }()
     
     // MARK: Init
@@ -133,248 +110,13 @@ class VoiceRecordingViewController: UIViewController, AVAudioRecorderDelegate {
     
     // MARK: Functions
     private func additionalSetup() {
-        view.backgroundColor = .secondarySystemBackground
-        setupButtonActions()
-        updateTimerLabel()
         title = "Gravação de Áudio"
+        view.backgroundColor = .secondarySystemBackground
+                
+        setupButtonActions()
+        recorder.updateTimerLabel()
         
         navigationController?.navigationBar.prefersLargeTitles = false
-    }
-    
-    private func startRecording() {
-        createURL()
-        
-        recordingState = .recording
-        
-        let settings: [String: Any] = [
-            AVFormatIDKey: Int(kAudioFormatMPEG4AAC),
-            AVSampleRateKey: 16000,
-            AVNumberOfChannelsKey: 1,
-            AVEncoderAudioQualityKey: AVAudioQuality.high.rawValue
-        ]
-            
-        do {
-            guard let audioURL else { return }
-            audioRecorder = try AVAudioRecorder(url: audioURL, settings: settings)
-            audioRecorder?.delegate = self
-            audioRecorder?.record()
-        } catch {
-            print("erro")
-        }
-            
-        // Atualiza o ícone do botão
-        updateRecordButtonIcon()
-            
-        timer = Timer.scheduledTimer(
-            timeInterval: 0.01,
-            target: self,
-            selector: #selector(updateTimer),
-            userInfo: nil,
-            repeats: true
-        )
-        startRecordingAnimation()
-    }
-        
-    private func pauseRecording() {
-        recordingState = .paused
-        updateRecordButtonIcon()
-        
-        timer?.invalidate()
-        timer = nil
-        stopRecordingAnimation()
-    }
-        
-    private func stopRecording() {
-            audioRecorder?.pause()
-            recordingState = .stopped
-            timer?.invalidate()
-            timer = nil
-            
-            // Atualiza o ícone do botão de gravação
-            updateRecordButtonIcon()
-            stopRecordingAnimation()
-        }
-        
-    private func updateRecordButtonIcon() {
-            let iconName: String
-            let iconColor: UIColor = .systemBackground
-            
-            switch recordingState {
-            case .stopped:
-                iconName = "stop.fill"
-            case .recording:
-                iconName = "pause.fill"
-            case .paused:
-                iconName = "play.fill"
-            }
-            
-            DispatchQueue.main.async { [weak self] in
-                guard let self = self else { return }
-                
-                let iconSize = self.recordButton.bounds.width * 0.4
-                let symbolConfig = UIImage.SymbolConfiguration(pointSize: iconSize, weight: .bold)
-                let iconImage = UIImage(systemName: iconName, withConfiguration: symbolConfig)
-                
-                self.recordButton.setImage(iconImage, for: .normal)
-                self.recordButton.tintColor = iconColor
-            }
-        }
-        
-    private func updateTimerLabel() {
-        let hours = Int(elapsedTime) / 3600
-        let minutes = (Int(elapsedTime) % 3600) / 60
-        let seconds = Int(elapsedTime) % 60
-            
-        DispatchQueue.main.async {
-            self.timerLabel.text = String(format: "%02d:%02d:%02d", hours, minutes, seconds)
-        }
-    }
-
-    private func resetRecording() {
-        stopRecording()
-        
-        elapsedTime = 0
-        
-        updateTimerLabel()
-    }
-    
-    private func showFinishConfirmationAlert() {
-        let alert = UIAlertController(
-            title: "Finalizar Gravação",
-            message: "Você deseja finalizar a gravação?",
-            preferredStyle: .alert
-        )
-        
-        let cancelAction = UIAlertAction(title: "Cancelar", style: .default)
-        let finishAction = UIAlertAction(title: "Finalizar", style: .cancel) { [weak self] _ in
-            self?.handleFinishRecording()
-        }
-        
-        finishAction.setValue(UIColor.systemGreen, forKey: "titleTextColor")
-        
-        alert.addAction(cancelAction)
-        alert.addAction(finishAction)
-        
-        present(alert, animated: true)
-    }
-    
-    private func showDeleteConfirmationAlert() {
-        let alert = UIAlertController(
-            title: "Deletar Áudio",
-            message: "Tem certeza de que deseja deletar esse áudio?",
-            preferredStyle: .alert
-        )
-        
-        let cancelAction = UIAlertAction(title: "Cancelar", style: .default)
-        let deleteAction = UIAlertAction(title: "Deletar", style: .cancel) { [weak self] _ in
-            self?.resetRecording()
-        }
-        
-        // Define a cor verde para o botão Iniciar Gravação
-        deleteAction.setValue(UIColor.systemRed, forKey: "titleTextColor")
-        
-        alert.addAction(cancelAction)
-        alert.addAction(deleteAction)
-        
-        present(alert, animated: true)
-    }
-        
-    private func handleFinishRecording() {
-        stopRecording()
-            
-        saveAudioRecording()
-            
-        resetRecording()
-        
-        changeScreen()
-    }
-    
-    private func changeScreen() {
-        let renameVC = RenameViewController()
-        renameVC.audioID = self.audioID
-        
-        let navController = UINavigationController(rootViewController: renameVC)
-
-        if let sheet = navController.sheetPresentationController {
-            sheet.detents = [.medium()]
-            sheet.prefersGrabberVisible = true
-        }
-
-        navController.modalPresentationStyle = .pageSheet
-        present(navController, animated: true)
-    }
-    
-    private func createURL() {
-        let audioSession = AVAudioSession.sharedInstance()
-        do {
-            try audioSession.setCategory(.playAndRecord, mode: .default, options: [.defaultToSpeaker])
-            try audioSession.setActive(true)
-            
-            let documentsPath = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask)[0]
-            audioURL = documentsPath.appendingPathComponent("audio_\(Date().timeIntervalSince1970).m4a")
-            print("Audio Salvo com o nome: \(String(describing: audioURL))")
-        } catch {
-            print("Erro ao iniciar gravação: \(error.localizedDescription)")
-        }
-    }
-        
-    /// Função reservada para implementação futura do salvamento do áudio
-    private func saveAudioRecording() {
-        guard let audioURL = audioURL else {
-            print("Erro ao criar URL para o áudio.")
-            return
-        }
-        let audio = AudioFileModel(id: UUID(), audioPath: audioURL.lastPathComponent)
-        audioID = audio.id
-            
-        Persistence.shared.createAudio(audio)
-    }
-    
-    ///whatafuck
-    func animateRecordingImage() {
-        UIView.animate(withDuration: 0.5, delay: 0, usingSpringWithDamping: 0.5, initialSpringVelocity: 1, options: .curveEaseInOut) {
-            self.recordingImage.transform = CGAffineTransform(scaleX: 1.2, y: 1.2)
-        } completion: { _ in
-            UIView.animate(withDuration: 0.5, delay: 0, usingSpringWithDamping: 0.5, initialSpringVelocity: 1, options: .curveEaseInOut) {
-                self.recordingImage.transform = .identity
-            }
-        }
-    }
-
-    // Inicia a animação
-    func startRecordingAnimation() {
-        currentImageIndex = 0
-        recordingImage.image = recordingImages[currentImageIndex]
-        
-        // Configura o timer para trocar as imagens a cada 0.5 segundos
-        recordingAnimationTimer = Timer.scheduledTimer(withTimeInterval: 0.6, repeats: true) { [weak self] _ in
-            guard let self = self else { return }
-            
-            self.currentImageIndex = (self.currentImageIndex + 1) % self.recordingImages.count
-            
-            UIView.transition(with: self.recordingImage,
-                              duration: 0.4,
-                              options: .transitionCrossDissolve,
-                              animations: {
-                                  self.recordingImage.image = self.recordingImages[self.currentImageIndex]
-                              },
-                              completion: nil)
-        }
-    }
-
-    // Para a animação
-    func stopRecordingAnimation() {
-        recordingAnimationTimer?.invalidate()
-        recordingAnimationTimer = nil
-        
-        // Volta para a imagem inicial (ou outra de sua escolha)
-        UIView.transition(with: recordingImage,
-                          duration: 0.3,
-                          options: .transitionCrossDissolve,
-                          animations: {
-            self.recordingImage.image = .emptyRec
-                          },
-                          completion: nil)
     }
 }
 
@@ -406,14 +148,18 @@ extension VoiceRecordingViewController: ViewCodeProtocol {
             buttonsStackView.centerXAnchor.constraint(equalTo: view.centerXAnchor),
             
             recordButton.widthAnchor.constraint(equalToConstant: 92),
-            recordButton.heightAnchor.constraint(equalToConstant: 92)
+            recordButton.heightAnchor.constraint(equalToConstant: 92),
+            
+            deleteAudioButton.widthAnchor.constraint(equalToConstant: 56),
+            deleteAudioButton.heightAnchor.constraint(equalToConstant: 56),
+            
+            finishedAudioButton.widthAnchor.constraint(equalToConstant: 56),
+            finishedAudioButton.heightAnchor.constraint(equalToConstant: 56)
         ])
     }
 }
 
 extension VoiceRecordingViewController {
-    
-    /// linkar botões com suas ações
     private func setupButtonActions() {
         recordButton.addTarget(self, action: #selector(recordButtonTapped), for: .touchUpInside)
         recordButton.addTarget(self, action: #selector(toggleRecordingAnimation), for: .touchDragEnter)
@@ -421,47 +167,50 @@ extension VoiceRecordingViewController {
         deleteAudioButton.addTarget(self, action: #selector(deleteButtonTapped), for: .touchUpInside)
     }
     
-    /// ações dos botões
-    @objc private func deleteButtonTapped() {
-        showDeleteConfirmationAlert()
+    @objc func deleteButtonTapped() {
+        recorder.showDeleteConfirmationAlert()
     }
        
-    @objc private func finishButtonTapped() {
-        showFinishConfirmationAlert()
+    @objc func finishButtonTapped() {
+        recorder.showFinishConfirmationAlert()
     }
        
-    @objc private func recordButtonTapped() {
-           switch recordingState {
-           case .stopped:
-               startRecording()
-               audioRecorder?.record()
-           case .paused:
-               audioRecorder?.record()
-               updateRecordButtonIcon()
+    @objc func recordButtonTapped() {
+        switch recorder.recordingState {
+        case .stopped:
+            recorder.startRecording()
+            recorder.audioRecorder?.record()
+        case .paused:
+            recorder.audioRecorder?.record()
+            recorder.updateRecordButtonIcon()
                
-               timer = Timer.scheduledTimer(
-                   timeInterval: 0.01,
-                   target: self,
-                   selector: #selector(updateTimer),
-                   userInfo: nil,
-                   repeats: true
-               )
-               startRecordingAnimation()
-           case .recording:
-               pauseRecording()
-               audioRecorder?.pause()
-           }
-       }
+            recorder.timer = Timer.scheduledTimer(
+                timeInterval: 0.01,
+                target: self,
+                selector: #selector(updateTimer),
+                userInfo: nil,
+                repeats: true
+            )
+            
+            recorder.startRecordingAnimation()
+        case .recording:
+            recorder.pauseRecording()
+            recorder.audioRecorder?.pause()
+            recorder.timer?.invalidate()
+            recorder.timer = nil
+
+        }
+    }
     
     // Função para iniciar/parar a animação
     @objc func toggleRecordingAnimation() {
-            if recordingState == .recording {
-                startRecordingAnimation()
+            if recorder.recordingState == .recording {
+                recorder.startRecordingAnimation()
             } else {
-                stopRecordingAnimation()
+                recorder.stopRecordingAnimation()
             }
-            audioRecorder?.stop()
-            audioRecorder = nil
+            recorder.audioRecorder?.stop()
+            recorder.audioRecorder = nil
             
             do {
                 try AVAudioSession.sharedInstance().setActive(false)
@@ -470,8 +219,8 @@ extension VoiceRecordingViewController {
             }
         }
     
-    @objc private func updateTimer() {
-        elapsedTime += 0.01
-        updateTimerLabel()
+    @objc func updateTimer() {
+        recorder.elapsedTime += 0.01
+        recorder.updateTimerLabel()
     }
 }
